Move movies api calls to LandingPage
