+++ src/components/pages/LandingPage.vue
@@ -7,6 +7,11 @@ import TheContainer from '@/components/common/TheContainer.vue';
 import WelcomeSection from '@/components/views/WelcomeSection.vue';
 import MovieCard from '@/components/views/MovieCard.vue';
 import MoviesSummary from '@/components/views/MoviesSummary.vue';
+import ScreeningsFilters from '../views/ScreeningsFilters.vue';
+import ScreeningsList from '../views/ScreeningsList.vue';
+import ScreeningsCard from '../views/ScreeningsCard.vue';
+import { getAllMovies } from '@/api/service/movies';
+
 
 export default defineComponent({
   components: {
@@ -17,7 +22,24 @@ export default defineComponent({
     WelcomeSection,
     MovieCard,
     MoviesSummary,
-}
+    ScreeningsFilters,
+    ScreeningsCard,
+    ScreeningsList,
+},
+   data() {
+    return {
+        movies: [],
+        filteredMovies: [],
+    };
+  },
+  async mounted() {
+    const response = await getAllMovies();
+    this.movies = response.data;
+    for (let i=0; i<3; i++) {
+       this.filteredMovies.push(this.movies[i])
+    }
+    console.log(this.movies)
+  }
 });
 </script>
 
@@ -28,8 +50,11 @@ export default defineComponent({
       <SignSection />
     </TheHeader>
     <WelcomeSection />
-    <MoviesSummary />
-    <ScreeningsFilters />
+    <TheContainer>
+      <MoviesSummary :movies="filteredMovies" />
+      <ScreeningsFilters />
+      <ScreeningsList :movies="movies" />
+    </TheContainer>
   </div>
 </template>
